Move log initialization out of init_config()
Since the log initialization makes a decision regarding whether it is
being called by the API or CLI and currently defaults to CLI mode, it is
not safe to be called to seed the API mode. By moving it out of the
init_config() call, we let API users initialize things with global state
and support plugins properly.